Trait iced_native::image::Renderer

pub trait Renderer: Renderer {
    type Handle: Clone + Hash;
    fn dimensions(&self, handle: &Self::Handle) -> (u32, u32);

    fn draw(&mut self, handle: Self::Handle, bounds: Rectangle);
}

A Renderer that can render raster graphics.

Associated Types

type Handle: Clone + Hash

The image Handle to be displayed. Iced exposes its own default implementation of a Handle

Required methods

fn dimensions(&self, handle: &Self::Handle) -> (u32, u32)

Returns the dimensions of an image for the given Handle.

fn draw(&mut self, handle: Self::Handle, bounds: Rectangle)

Draws an image with the given Handle and inside the provided bounds.