Struct iced_native::overlay::menu::Menu

pub struct Menu<'a, T, Renderer>where
    Renderer: Renderer,
    Renderer::Theme: StyleSheet,{ /* private fields */ }

A list of selectable options.

Implementations

impl<'a, T, Renderer> Menu<'a, T, Renderer>where T: ToString + Clone, Renderer: Renderer + 'a, Renderer::Theme: StyleSheet + StyleSheet + StyleSheet,

pub fn new( state: &'a mut State, options: &'a [T], hovered_option: &'a mut Option<usize>, last_selection: &'a mut Option<T> ) -> Self

Creates a new Menu with the given State, a list of options, and the message to produced when an option is selected.

pub fn width(self, width: u16) -> Self

Sets the width of the Menu.

pub fn padding<P: Into<Padding>>(self, padding: P) -> Self

Sets the Padding of the Menu.

pub fn text_size(self, text_size: u16) -> Self

Sets the text size of the Menu.

pub fn font(self, font: Renderer::Font) -> Self

Sets the font of the Menu.

pub fn style( self, style: impl Into<<Renderer::Theme as StyleSheet>::Style> ) -> Self

Sets the style of the Menu.

pub fn overlay<Message: 'a>( self, position: Point, target_height: f32 ) -> Element<'a, Message, Renderer>

Turns the Menu into an overlay Element at the given target position.

The target_height will be used to display the menu either on top of the target or under it, depending on the screen position and the dimensions of the Menu.