Struct iced_native::overlay::Element

pub struct Element<'a, Message, Renderer> { /* private fields */ }

A generic Overlay.

Implementations

impl<'a, Message, Renderer> Element<'a, Message, Renderer>where Renderer: Renderer,

pub fn new( position: Point, overlay: Box<dyn Overlay<Message, Renderer> + 'a> ) -> Self

Creates a new Element containing the given Overlay.

pub fn position(&self) -> Point

Returns the position of the Element.

pub fn translate(self, translation: Vector) -> Self

Translates the Element.

pub fn map<B>(self, f: &'a dyn Fn(Message) -> B) -> Element<'a, B, Renderer>where Message: 'a, Renderer: 'a, B: 'a,

Applies a transformation to the produced message of the Element.

pub fn layout( &self, renderer: &Renderer, bounds: Size, translation: Vector ) -> Node

Computes the layout of the Element in the given bounds.

pub fn on_event( &mut self, event: Event, layout: Layout<'_>, cursor_position: Point, renderer: &Renderer, clipboard: &mut dyn Clipboard, shell: &mut Shell<'_, Message> ) -> Status

Processes a runtime Event.

pub fn mouse_interaction( &self, layout: Layout<'_>, cursor_position: Point, viewport: &Rectangle, renderer: &Renderer ) -> Interaction

Returns the current mouse::Interaction of the Element.

pub fn draw( &self, renderer: &mut Renderer, theme: &Renderer::Theme, style: &Style, layout: Layout<'_>, cursor_position: Point )

Draws the Element and its children using the given Layout.

pub fn operate( &mut self, layout: Layout<'_>, renderer: &Renderer, operation: &mut dyn Operation<Message> )

Applies a widget::Operation to the Element.

pub fn is_over(&self, layout: Layout<'_>, cursor_position: Point) -> bool

Returns true if the cursor is over the Element.