Struct iced_native::Padding

pub struct Padding {
    pub top: u16,
    pub right: u16,
    pub bottom: u16,
    pub left: u16,
}

An amount of space to pad for each side of a box

You can leverage the From trait to build Padding conveniently:

let padding = Padding::from(20);              // 20px on all sides
let padding = Padding::from([10, 20]);        // top/bottom, left/right
let padding = Padding::from([5, 10, 15, 20]); // top, right, bottom, left

Normally, the padding method of a widget will ask for an Into<Padding>, so you can easily write:

impl Widget {
    pub fn padding(mut self, padding: impl Into<Padding>) -> Self {
        // ...
        self
    }
}

let widget = Widget::new().padding(20);              // 20px on all sides
let widget = Widget::new().padding([10, 20]);        // top/bottom, left/right
let widget = Widget::new().padding([5, 10, 15, 20]); // top, right, bottom, left

Fields
top: u16

Top padding

right: u16

Right padding

bottom: u16

Bottom padding

left: u16

Left padding

Implementations

impl Padding

pub const ZERO: Padding

Padding of zero

pub const fn new(padding: u16) -> Padding

Create a Padding that is equal on all sides

pub fn vertical(self) -> u16

Returns the total amount of vertical Padding.

pub fn horizontal(self) -> u16

Returns the total amount of horizontal Padding.