Struct iced_native::Rectangle

pub struct Rectangle<T = f32> {
    pub x: T,
    pub y: T,
    pub width: T,
    pub height: T,
}

A rectangle.

Fields
x: T

X coordinate of the top-left corner.

y: T

Y coordinate of the top-left corner.

width: T

Width of the rectangle.

height: T

Height of the rectangle.

Implementations

impl Rectangle<f32>

pub fn new(top_left: Point, size: Size<f32>) -> Rectangle<f32>

Creates a new Rectangle with its top-left corner in the given Point and with the provided Size.

pub fn with_size(size: Size<f32>) -> Rectangle<f32>

Creates a new Rectangle with its top-left corner at the origin and with the provided Size.

pub fn center(&self) -> Point

Returns the Point at the center of the Rectangle.

pub fn center_x(&self) -> f32

Returns the X coordinate of the Point at the center of the Rectangle.

pub fn center_y(&self) -> f32

Returns the Y coordinate of the Point at the center of the Rectangle.

pub fn position(&self) -> Point

Returns the position of the top left corner of the Rectangle.

pub fn size(&self) -> Size<f32>

Returns the Size of the Rectangle.

pub fn contains(&self, point: Point) -> bool

Returns true if the given Point is contained in the Rectangle.

pub fn intersection(&self, other: &Rectangle<f32>) -> Option<Rectangle<f32>>

Computes the intersection with the given Rectangle.

pub fn snap(self) -> Rectangle<u32>

Snaps the Rectangle to unsigned integer coordinates.