Module iced_native::subscription

Listen to external events in your application.

Traits

Recipe

The description of a Subscription.

Functions

events

Returns a Subscription to all the runtime events.

events_with

Returns a Subscription that filters all the runtime events with the provided function, producing messages accordingly.

run

Returns a Subscription that will create and asynchronously run the given [Stream].

unfold

Returns a Subscription that will create and asynchronously run a [Stream] that will call the provided closure to produce every Message.

Type Definitions

EventStream

A stream of runtime events.

Subscription

A request to listen to external events.

Tracker

A native Subscription tracker.