Function iced_native::subscription::events_with

pub fn events_with<Message>(
    f: fn(_: Event, _: Status) -> Option<Message>
) -> Subscription<Message> where
    Message: 'static + MaybeSend, 

Returns a Subscription that filters all the runtime events with the provided function, producing messages accordingly.

This subscription will call the provided function for every Event handled by the runtime. If the function:

Returns None, the Event will be discarded.

Returns Some message, the Message will be produced.