Module iced_native::widget::button

Allow your users to perform actions by pressing a button.

A Button has some local State.

Structs

Button

A generic widget that produces a message when pressed.

State

The local state of a Button.

Style

The appearance of a button.

Traits

StyleSheet

A set of rules that dictate the style of a button.

Functions

draw

Draws a Button.

layout

Computes the layout of a Button.

mouse_interaction

Returns the mouse::Interaction of a Button.

update

Processes the given Event and updates the State of a Button accordingly.