Struct iced_native::widget::column::Column

pub struct Column<'a, Message, Renderer> { /* private fields */ }

A container that distributes its contents vertically.

Implementations

impl<'a, Message, Renderer> Column<'a, Message, Renderer>

pub fn new() -> Self

Creates an empty Column.

pub fn with_children(children: Vec<Element<'a, Message, Renderer>>) -> Self

Creates a Column with the given elements.

pub fn spacing(self, amount: impl Into<Pixels>) -> Self

Sets the vertical spacing between elements.

Custom margins per element do not exist in iced. You should use this method instead! While less flexible, it helps you keep spacing between elements consistent.

pub fn padding<P: Into<Padding>>(self, padding: P) -> Self

Sets the Padding of the Column.

pub fn width(self, width: impl Into<Length>) -> Self

Sets the width of the Column.

pub fn height(self, height: impl Into<Length>) -> Self

Sets the height of the Column.

pub fn max_width(self, max_width: impl Into<Pixels>) -> Self

Sets the maximum width of the Column.

pub fn align_items(self, align: Alignment) -> Self

Sets the horizontal alignment of the contents of the Column .

pub fn push(self, child: impl Into<Element<'a, Message, Renderer>>) -> Self

Adds an element to the Column.