Module iced_native::widget::helpers

Helper functions to create pure widgets.

Functions

button

Creates a new Button with the provided content.

checkbox

Creates a new Checkbox.

column

Creates a new Column with the given children.

container

Creates a new Container with the provided content.

horizontal_rule

Creates a horizontal Rule with the given height.

horizontal_space

Creates a new horizontal Space with the given Length.

image

Creates a new Image.

pick_list

Creates a new PickList.

progress_bar

Creates a new ProgressBar.

radio

Creates a new Radio.

row

Creates a new Row with the given children.

scrollable

Creates a new Scrollable with the provided content.

slider

Creates a new Slider.

svg

Creates a new Svg widget from the given Handle.

text

Creates a new Text widget with the provided content.

text_input

Creates a new TextInput.

toggler

Creates a new Toggler.

tooltip

Creates a new Tooltip with the provided content, tooltip text, and tooltip::Position.

vertical_rule

Creates a vertical Rule with the given width.

vertical_space

Creates a new vertical Space with the given Length.