Module iced_native::widget::image::viewer

Zoom and pan on an image.

Structs

State

The local state of a Viewer.

Viewer

A frame that displays an image with the ability to zoom in/out and pan.

Functions

image_size

Returns the bounds of the underlying image, given the bounds of the Viewer. Scaling will be applied and original aspect ratio will be respected.