Module iced_native::widget::operation

Query or update internal widget state.

Re-exports

pub use focusable::Focusable;

pub use scrollable::Scrollable;

Modules

focusable

Operate on widgets that can be focused.

scrollable

Operate on widgets that can be scrolled.

Enums

Outcome

The result of an Operation.

Traits

Operation

A piece of logic that can traverse the widget tree of an application in order to query or update some widget state.