Module iced_native::widget::operation::scrollable

Operate on widgets that can be scrolled.

Traits

Scrollable

The internal state of a widget that can be scrolled.

Functions

snap_to

Produces an Operation that snaps the widget with the given Id to the provided percentage.