Trait iced_native::widget::operation::scrollable::Scrollable

pub trait Scrollable {
    // Required method
    fn snap_to(&mut self, offset: RelativeOffset);
}

The internal state of a widget that can be scrolled.

Required Methods

fn snap_to(&mut self, offset: RelativeOffset)

Snaps the scroll of the widget to the given percentage along the horizontal & vertical axis.