Trait iced_native::widget::pane_grid::StyleSheet

pub trait StyleSheet {
    type Style: Default;

    // Required methods
    fn picked_split(&self, style: &Self::Style) -> Option<Line>;
    fn hovered_split(&self, style: &Self::Style) -> Option<Line>;
}

A set of rules that dictate the style of a container.

Required Associated Types

type Style: Default

The supported style of the StyleSheet.

Required Methods

fn picked_split(&self, style: &Self::Style) -> Option<Line>

The Line to draw when a split is picked.

fn hovered_split(&self, style: &Self::Style) -> Option<Line>

The Line to draw when a split is hovered.