Module iced_native::widget::scrollable

Navigate an endless amount of content with a scrollbar.

Modules

style

The styles of a Scrollable.

Structs

Scrollable

A widget that can vertically display an infinite amount of content with a scrollbar.

State

The local state of a Scrollable.

Traits

StyleSheet

A set of rules that dictate the style of a scrollable.

Functions

draw

Draws a Scrollable.

layout

Computes the layout of a Scrollable.

mouse_interaction

Computes the current mouse::Interaction of a Scrollable.

update

Processes an Event and updates the State of a Scrollable accordingly.