Struct iced_native::widget::scrollable::Scrollable

pub struct Scrollable<'a, Message, Renderer> { /* fields omitted */ }

A widget that can vertically display an infinite amount of content with a scrollbar.

Implementations

impl<'a, Message, Renderer: Renderer> Scrollable<'a, Message, Renderer>

pub fn new(state: &'a mut State) -> Self

Creates a new Scrollable with the given State.

pub fn spacing(self, units: u16) -> Self

Sets the vertical spacing between elements.

Custom margins per element do not exist in Iced. You should use this method instead! While less flexible, it helps you keep spacing between elements consistent.

pub fn padding<P: Into<Padding>>(self, padding: P) -> Self

Sets the Padding of the Scrollable.

pub fn width(self, width: Length) -> Self

Sets the width of the Scrollable.

pub fn height(self, height: Length) -> Self

Sets the height of the Scrollable.

pub fn max_width(self, max_width: u32) -> Self

Sets the maximum width of the Scrollable.

pub fn max_height(self, max_height: u32) -> Self

Sets the maximum height of the Scrollable in pixels.

pub fn align_items(self, align_items: Alignment) -> Self

Sets the horizontal alignment of the contents of the Scrollable .

pub fn scrollbar_width(self, scrollbar_width: u16) -> Self

Sets the scrollbar width of the Scrollable . Silently enforces a minimum value of 1.

pub fn scrollbar_margin(self, scrollbar_margin: u16) -> Self

Sets the scrollbar margin of the Scrollable .

pub fn scroller_width(self, scroller_width: u16) -> Self

Sets the scroller width of the Scrollable .

It silently enforces a minimum value of 1.

pub fn on_scroll(self, f: impl Fn(f32) -> Message + 'static) -> Self

Sets a function to call when the Scrollable is scrolled.

The function takes the new relative offset of the Scrollable (e.g. 0 means top, while 1 means bottom).

pub fn style(self, style_sheet: impl Into<Box<dyn StyleSheet + 'a>>) -> Self

Sets the style of the Scrollable .

pub fn push<E>(self, child: E) -> Self where
E: Into<Element<'a, Message, Renderer>>,

Adds an element to the Scrollable.