Module iced_native::widget::text_input

Display fields that can be filled with text.

A TextInput has some local State.

Re-exports

pub use cursor::Cursor;

Modules

cursor

Track the cursor of a text input.

Structs

Appearance

The appearance of a text input.

Id

The identifier of a TextInput.

State

The state of a TextInput.

TextInput

A field that can be filled with text.

Value

The value of a TextInput.

Traits

StyleSheet

A set of rules that dictate the style of a text input.

Functions

draw

Draws the TextInput with the given Renderer, overriding its Value if provided.

focus

Produces a Command that focuses the TextInput with the given Id.

layout

Computes the layout of a TextInput.

mouse_interaction

Computes the current mouse::Interaction of the TextInput.

update

Processes an Event and updates the State of a TextInput accordingly.