Struct iced_native::widget::toggler::Toggler

pub struct Toggler<'a, Message, Renderer>where
    Renderer: Renderer,
    Renderer::Theme: StyleSheet,{ /* private fields */ }

A toggler widget.

Example

pub enum Message {
    TogglerToggled(bool),
}

let is_active = true;

Toggler::new(is_active, String::from("Toggle me!"), |b| Message::TogglerToggled(b));

Implementations

impl<'a, Message, Renderer> Toggler<'a, Message, Renderer>where
Renderer: Renderer,
Renderer::Theme: StyleSheet,

pub const DEFAULT_SIZE: u16 = 20u16

The default size of a Toggler.

pub fn new<F>(is_active: bool, label: impl Into<Option<String>>, f: F) -> Selfwhere
F: 'a + Fn(bool) -> Message,

Creates a new Toggler.

It expects:

a boolean describing whether the Toggler is checked or not

An optional label for the Toggler

a function that will be called when the Toggler is toggled. It will receive the new state of the Toggler and must produce a Message.

pub fn size(self, size: u16) -> Self

Sets the size of the Toggler.

pub fn width(self, width: Length) -> Self

Sets the width of the Toggler.

pub fn text_size(self, text_size: u16) -> Self

Sets the text size o the Toggler.

pub fn text_alignment(self, alignment: Horizontal) -> Self

Sets the horizontal alignment of the text of the Toggler

pub fn spacing(self, spacing: u16) -> Self

Sets the spacing between the Toggler and the text.

pub fn font(self, font: Renderer::Font) -> Self

Sets the Font of the text of the Toggler

pub fn style(
self,
style: impl Into<<Renderer::Theme as StyleSheet>::Style>
) -> Self

Sets the style of the Toggler.