Struct iced_native::widget::tree::Tree

pub struct Tree {
    pub tag: Tag,
    pub state: State,
    pub children: Vec<Tree>,
}

A persistent state widget tree.

A Tree is normally associated with a specific widget in the widget tree.

Fields
tag: Tag

The tag of the Tree.

state: State

The State of the Tree.

children: Vec<Tree>

The children of the root widget of the Tree.

Implementations

impl Tree

pub fn empty() -> Self

Creates an empty, stateless Tree with no children.

pub fn new<'a, Message, Renderer>(
widget: impl Borrow<dyn Widget<Message, Renderer> + 'a>
) -> Selfwhere
Renderer: Renderer,

Creates a new Tree for the provided [Element].

pub fn diff<'a, Message, Renderer>(
&mut self,
new: impl Borrow<dyn Widget<Message, Renderer> + 'a>
)where
Renderer: Renderer,

Reconciliates the current tree with the provided [Element].

If the tag of the [Element] matches the tag of the Tree, then the [Element] proceeds with the reconciliation (i.e. Widget::diff is called).

Otherwise, the whole Tree is recreated.

pub fn diff_children<'a, Message, Renderer>(
&mut self,
new_children: &[impl Borrow<dyn Widget<Message, Renderer> + 'a>]
)where
Renderer: Renderer,

Reconciliates the children of the tree with the provided list of [Element].

pub fn diff_children_custom<T>(
&mut self,
new_children: &[T],
diff: impl Fn(&mut Tree, &T),
new_state: impl Fn(&T) -> Self
)

Reconciliates the children of the tree with the provided list of [Element] using custom logic both for diffing and creating new widget state.