Enum iced_native::window::Action

pub enum Action<T> {
    Close,
    Drag,
    Resize {
        width: u32,
        height: u32,
    },
    Maximize(bool),
    Minimize(bool),
    Move {
        x: i32,
        y: i32,
    },
    ChangeMode(Mode),
    FetchMode(Box<dyn FnOnce(Mode) -> T + 'static>),
    ToggleMaximize,
    ToggleDecorations,
    RequestUserAttention(Option<UserAttention>),
    GainFocus,
}

An operation to be performed on some window.

Variants

Close

Closes the current window and exits the application.

Drag

Moves the window with the left mouse button until the button is released.

There’s no guarantee that this will work unless the left mouse button was pressed immediately before this function is called.

Resize

Fields

width: u32

The new logical width of the window

height: u32

The new logical height of the window

Resize the window.

Maximize(bool)

Sets the window to maximized or back

Minimize(bool)

Set the window to minimized or back

Move

Fields

x: i32

The new logical x location of the window

y: i32

The new logical y location of the window

Move the window.

Unsupported on Wayland.

ChangeMode(Mode)

Change the Mode of the window.

FetchMode(Box<dyn FnOnce(Mode) -> T + 'static>)

Fetch the current Mode of the window.

ToggleMaximize

Toggle the window to maximized or back

ToggleDecorations

Toggle whether window has decorations.

Platform-specific

X11: Not implemented.

Web: Unsupported.

RequestUserAttention(Option<UserAttention>)

Request user attention to the window, this has no effect if the application is already focused. How requesting for user attention manifests is platform dependent, see UserAttention for details.

Providing None will unset the request for user attention. Unsetting the request for user attention might not be done automatically by the WM when the window receives input.

Platform-specific

iOS / Android / Web: Unsupported.

macOS: None has no effect.

X11: Requests for user attention must be manually cleared.

Wayland: Requires xdg_activation_v1 protocol, None has no effect.

GainFocus

Bring the window to the front and sets input focus. Has no effect if the window is already in focus, minimized, or not visible.

This method steals input focus from other applications. Do not use this method unless you are certain that’s what the user wants. Focus stealing can cause an extremely disruptive user experience.

Platform-specific

Web / Wayland: Unsupported.

Implementations

impl<T> Action<T>

pub fn map<A>(self, f: impl Fn(T) -> A + 'static + MaybeSend + Sync) -> Action<A>where T: 'static,

Maps the output of a window Action using the provided closure.