Module iced_native::widget

Use the built-in widgets or create your own.

Built-in widgets

Every built-in drawable widget has its own module with a Renderer trait that must be implemented by a renderer before being able to use it as a Widget.

Custom widgets

If you want to implement a custom widget, you simply need to implement the Widget trait. You can use the API of the built-in widgets as a guide or source of inspiration.

Re-exports

pub use button::Button;

pub use checkbox::Checkbox;

pub use column::Column;

pub use container::Container;

pub use image::Image;

pub use pane_grid::PaneGrid;

pub use pick_list::PickList;

pub use progress_bar::ProgressBar;

pub use radio::Radio;

pub use row::Row;

pub use rule::Rule;

pub use scrollable::Scrollable;

pub use slider::Slider;

pub use space::Space;

pub use svg::Svg;

pub use text::Text;

pub use text_input::TextInput;

pub use toggler::Toggler;

pub use tooltip::Tooltip;

pub use tree::Tree;

pub use operation::Operation;

pub use helpers::*;

Modules

button

Allow your users to perform actions by pressing a button.

checkbox

Show toggle controls using checkboxes.

column

Distribute content vertically.

container

Decorate content and apply alignment.

helpers

Helper functions to create pure widgets.

image

Display images in your user interface.

operation

Query or update internal widget state.

pane_grid

Let your users split regions of your application and organize layout dynamically.

pick_list

Display a dropdown list of selectable values.

progress_bar

Provide progress feedback to your users.

radio

Create choices using radio buttons.

row

Distribute content horizontally.

rule

Display a horizontal or vertical rule for dividing content.

scrollable

Navigate an endless amount of content with a scrollbar.

slider

Display an interactive selector of a single value from a range of values.

space

Distribute content vertically.

svg

Display vector graphics in your application.

text

Write some text for your users to read.

text_input

Display fields that can be filled with text.

toggler

Show toggle controls using togglers.

tooltip

Display a widget over another.

tree

Store internal widget state in a state tree to ensure continuity.

Structs

Action

An operation to be performed on the widget tree.

Id

The identifier of a generic widget.

Traits

Widget

A component that displays information and allows interaction.