Function iced_native::widget::text::draw

pub fn draw<Renderer>(
    renderer: &mut Renderer, 
    style: &Style, 
    layout: Layout<'_>, 
    content: &str, 
    font: Renderer::Font, 
    size: Option<u16>, 
    color: Option<Color>, 
    horizontal_alignment: Horizontal, 
    vertical_alignment: Vertical
) where
    Renderer: Renderer, 

Draws text using the same logic as the Text widget.

Specifically:

If no size is provided, the default text size of the Renderer will be used.

If no color is provided, the renderer::Style::text_color will be used.

The alignment attributes do not affect the position of the bounds of the Layout.